Move Ethercat manager class in a separate library
