+++ device/CiA402MotionControl/CMakeLists.txt
@@ -5,10 +5,10 @@
 add_yarp_device(
     NAME CiA402MotionControl
     TYPE yarp::dev::CiA402MotionControl
-    SOURCES CiA402MotionControl.cpp CiA402StateMachine.cpp EthercatManager.cpp
-    PUBLIC_HEADERS CiA402MotionControl.h CiA402StateMachine.h EthercatManager.h LogComponent.h
+    SOURCES CiA402MotionControl.cpp CiA402StateMachine.cpp
+    PUBLIC_HEADERS CiA402MotionControl.h CiA402StateMachine.h
     PUBLIC_LINK_LIBRARIES
       YARP::YARP_os
       YARP::YARP_dev
-      soem
+      CiA402::CiA402
     CONFIGURE_PACKAGE_NAME CiA402MotionControl)
