+++ device/CiA402MotionControl/CiA402StateMachine.cpp
