+++ src/CiA402/include/CiA402/LogComponent.h
